update back to binned

+++ fermi/monte_carlo/simps/code/simulate.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-from lande.fermi.likelihood.roi_gtlike import Gtlike
+from lande.fermi.likelihood.roi_gtlike import Gtlike, UnbinnedGtlike
 
 from argparse import ArgumentParser
 import shutil
@@ -17,7 +17,6 @@ from uw.like.pointspec_helpers import PointSource
 from uw.like.Models import PowerLaw
 from uw.like.roi_state import PointlikeState
 
-from lande.fermi.likelihood.diffuse import get_sreekumar
 from lande.fermi.likelihood.save import sourcedict
 from lande.fermi.data.catalogs import dict2fgl
 from lande.fermi.data.livetime import gtltcube
@@ -29,9 +28,11 @@ parser = ArgumentParser()
 parser.add_argument("i", type=int)
 parser.add_argument("--time", required=True, choices=['2fgl', '1day', '2years'])
 parser.add_argument("--flux", required=True, type=float)
-parser.add_argument("--position", required=True, choices=['galcenter', 'allsky' ])
+parser.add_argument("--position", required=True, choices=['galcenter', 'allsky', 'bad', 'pole' ])
 parser.add_argument("--emin", required=True, type=float)
 parser.add_argument("--emax", required=True, type=float)
+parser.add_argument("--phibins", required=True, type=int)
+parser.add_argument("--savedata", default=False, action='store_true')
 
 args= parser.parse_args()
 
@@ -47,19 +48,26 @@ position=args.position
 time=args.time
 emin=args.emin
 emax=args.emax
+phibins=args.phibins
 
 if position == 'galcenter':
     roi_dir = SkyDir(0,0,SkyDir.GALACTIC)
 elif position == 'allsky':
     roi_dir=random_on_sphere()
+elif position == 'bad':
+    roi_dir=SkyDir(314.4346,-69.5670,SkyDir.GALACTIC)
+elif position == 'pole':
+    roi_dir=SkyDir(0,-90,SkyDir.GALACTIC)
 
 model_mc = PowerLaw(index=2)
 model_mc.set_flux(flux, emin=emin, emax=emax)
 
 ps = PointSource(name=name, model=model_mc, skydir=roi_dir)
-sreekumar = get_sreekumar()
 
-tempdir=mkdtemp(prefix='/scratch/')
+if args.savedata:
+    tempdir='savedir'
+else:
+    tempdir=mkdtemp(prefix='/scratch/')
 
 if time == '1day':
     ft2 = join(tempdir,'ft2.fits')
@@ -88,18 +96,20 @@ if not exists(ft1):
     mc = MonteCarlo(
         ft1=ft1,
         ft2=ft2,
-        sources=[ps,sreekumar],
+        sources=[ps],
         emin=emin,
         emax=emax,
         irf=irf,
         maxROI=roi_size,
         savedir=tempdir,
         seed=i,
+        mc_energy=True,
         **mc_kwargs)
     mc.simulate()
 
 if not exists(ltcube):
-    gtltcube(evfile=ft1, scfile=ft2, outfile=ltcube, dcostheta=0.025, binsz=1)
+    gtltcube(evfile=ft1, scfile=ft2, outfile=ltcube, dcostheta=0.025, binsz=1, phibins=phibins)
+    #gtltcube(evfile=ft1, scfile=ft2, outfile=ltcube, dcostheta=0.0125, binsz=0.25, phibins=phibins)
 
 ds = DataSpecification(
     ft1files = ft1,
@@ -121,7 +131,7 @@ sa = SpectralAnalysis(ds,
 
 roi = sa.roi(roi_dir=roi_dir, 
              point_sources=[ps],
-             diffuse_sources=[sreekumar],
+             diffuse_sources=None,
             )
 state = PointlikeState(roi)
 
@@ -131,32 +141,34 @@ results = dict(
     time = time,
     emin = emin,
     emax = emax,
-    istr = istr)
+    istr = istr,
+    phibins = phibins)
 
-mc=sourcedict(roi, name)
+mc=sourcedict(roi, name, save_TS=False)
 
 roi.print_summary()
 roi.fit(use_gradient=False)
 roi.print_summary()
 
-fit=sourcedict(roi, name)
+fit=sourcedict(roi, name, save_TS=False)
 
 results['pointlike'] = dict(mc=mc, fit=fit)
 
 state.restore()
 
-gtlike = Gtlike(roi, binsz=1/8., bigger_roi=False, 
-                enable_edisp=True, fix_pointlike_ltcube=True)
+#gtlike = UnbinnedGtlike(roi, savedir=tempdir)
+gtlike = Gtlike(roi, savedir=tempdir)
 like = gtlike.like
 
-mc=sourcedict(like, name)
+mc=sourcedict(like, name, save_TS=False)
 
 like.fit(covar=True)
 
-fit = sourcedict(like, name)
+fit = sourcedict(like, name, save_TS=False)
 
 results['gtlike'] = dict(mc=mc, fit=fit)
 
 savedict('results_%s.yaml' % istr, results)
 
-shutil.rmtree(tempdir)
+if not args.savedata:
+    shutil.rmtree(tempdir)
